Move client creation form into CreateClient directory
