+++ web-app/src/components/CreateClient/CreateClientForm/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { PropTypes as PT } from 'prop-types';
 import container from './container';
-import employeeStatus from '../../utilities/employeeStatus';
-import { Form, Input } from '../common';
+import employeeStatus from '../../../utilities/employeeStatus';
+import { Form, Input } from '../../common';
 import { FormContainer } from './styled';
 
 export const CreateClientForm = props => {
